+++ packages/web3-eth-personal/src/personal.ts
@@ -16,8 +16,8 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import { Web3Context } from 'web3-core';
-import { Address, HexString, Transaction } from 'web3-types';
-import { EthPersonalAPI } from './eth_personal_api';
+import { Address, EthPersonalAPI, HexString, Transaction } from 'web3-types';
+
 import * as rpcWrappers from './rpc_method_wrappers';
 
 export class Personal extends Web3Context<EthPersonalAPI> {