+++ packages/web3-eth/src/utils/transaction_builder.ts
@@ -31,18 +31,20 @@ import {
 	Hardfork,
 	Transaction,
 	TransactionWithLocalWalletIndex,
+	Common,
+	Web3NetAPI,
 } from 'web3-types';
 import { Web3Context } from 'web3-core';
 import { privateKeyToAddress } from 'web3-eth-accounts';
-import { getId, Web3NetAPI } from 'web3-net';
+import { getId } from 'web3-net';
 import { isNullish, isNumber } from 'web3-validator';
-import { NUMBER_DATA_FORMAT } from '../constants';
 import {
 	InvalidTransactionWithSender,
 	LocalWalletNotAvailableError,
 	TransactionDataAndInputError,
 	UnableToPopulateNonceError,
-} from '../errors';
+} from 'web3-errors';
+import { NUMBER_DATA_FORMAT } from '../constants';
 // eslint-disable-next-line import/no-cycle
 import { getChainId, getTransactionCount } from '../rpc_method_wrappers';
 import { detectTransactionType } from './detect_transaction_type';
@@ -161,6 +163,17 @@ export async function defaultTransactionBuilder<ReturnType = Record<string, unkn
 	}
 
 	if (isNullish(populatedTransaction.common)) {
+		if (options.web3Context.defaultCommon) {
+			const common = options.web3Context.defaultCommon as unknown as Common;
+			const chainId = common.customChain.chainId as string;
+			const networkId = common.customChain.networkId as string;
+			const name = common.customChain.name as string;
+			populatedTransaction.common = {
+				...common,
+				customChain: { chainId, networkId, name },
+			};
+		}
+
 		if (isNullish(populatedTransaction.chain)) {
 			populatedTransaction.chain = options.web3Context.defaultChain as ValidChains;
 		}