+++ packages/web3-eth/test/integration/web3_eth/send_signed_transaction.test.ts
@@ -290,8 +290,7 @@ describe('Web3Eth.sendSignedTransaction', () => {
 				gas: '0x5208',
 			});
 
-			// TODO: Debug why the assertions are not being called
-			// expect.assertions(1);
+			expect.assertions(1);
 		});
 	});
 });