+++ packages/web3-eth/test/unit/rpc_method_wrappers/get_logs.test.ts
@@ -16,14 +16,14 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 import { Web3Context } from 'web3-core';
 import { DEFAULT_RETURN_FORMAT, FMT_BYTES, FMT_NUMBER, format } from 'web3-utils';
+import { Web3EthExecutionAPI } from 'web3-types';
+import { ethRpcMethods } from 'web3-rpc-methods';
 
-import { getLogs as rpcMethodsGetLogs } from '../../../src/rpc_methods';
-import { Web3EthExecutionAPI } from '../../../src/web3_eth_execution_api';
 import { getLogs } from '../../../src/rpc_method_wrappers';
 import { mockRpcResponse, testData } from './fixtures/get_logs';
 import { logSchema } from '../../../src/schemas';
 
-jest.mock('../../../src/rpc_methods');
+jest.mock('web3-rpc-methods');
 
 describe('getLogs', () => {
 	let web3Context: Web3Context<Web3EthExecutionAPI>;
@@ -35,9 +35,9 @@ describe('getLogs', () => {
 	it.each(testData)(
 		`should call rpcMethods.getLogs with expected parameters\nTitle: %s\nInput parameters: %s\n`,
 		async (_, inputParameters) => {
-			(rpcMethodsGetLogs as jest.Mock).mockResolvedValueOnce(mockRpcResponse);
+			(ethRpcMethods.getLogs as jest.Mock).mockResolvedValueOnce(mockRpcResponse);
 			await getLogs(web3Context, ...inputParameters, DEFAULT_RETURN_FORMAT);
-			expect(rpcMethodsGetLogs).toHaveBeenCalledWith(
+			expect(ethRpcMethods.getLogs).toHaveBeenCalledWith(
 				web3Context.requestManager,
 				...inputParameters,
 			);
@@ -55,7 +55,7 @@ describe('getLogs', () => {
 
 				return format(logSchema, res, expectedReturnFormat);
 			});
-			(rpcMethodsGetLogs as jest.Mock).mockResolvedValueOnce(mockRpcResponse);
+			(ethRpcMethods.getLogs as jest.Mock).mockResolvedValueOnce(mockRpcResponse);
 
 			const result = await getLogs(web3Context, ...inputParameters, expectedReturnFormat);
 			expect(result).toStrictEqual(expectedFormattedResult);